Refactor imports and components in various files
